fix: Ensure overlay invalidation occurs
