+++ src/Uno.UI.Runtime.Skia.Gtk/GLRenderSurfaceBase.cs
@@ -54,8 +54,6 @@ namespace Uno.UI.Runtime.Skia
 			WUX.Window.InvalidateRender
 				+= () =>
 				{
-					// TODO Uno: Make this invalidation less often if possible.
-					InvalidateOverlays();
 					QueueRender();
 				};
